Refactored, message event out of the index.js
